+++ src/appleseed/renderer/modeling/environmentshader/backgroundenvironmentshader.cpp
@@ -88,7 +88,7 @@ namespace
         {
             const Vector2d& s = pixel_context.get_sample_position();
             const InputValues* input_values =
-                input_evaluator.evaluate<InputValues>(m_inputs, Vector2d(s[0], 1.0 - s[1]));
+                input_evaluator.evaluate<InputValues>(m_inputs, Vector2f(s[0], 1.0 - s[1]));
 
             shading_result.m_color_space = ColorSpaceSpectral;
             shading_result.m_main.m_color = input_values->m_color;
